создание входа и регистрации

+++ notes.controller.js
@@ -1,12 +1,8 @@
 const chalk = require("chalk");
 const Note = require("./models/Note");
 
-async function saveNotes(notes) {
-  await fs.writeFile(notesPath, JSON.stringify(notes));
-}
-
-async function addNote(title) {
-  await Note.create({ title });
+async function addNote(title, owner) {
+  await Note.create({ title, owner });
 
   console.log(chalk.green.inverse("Node was added!"));
 }
@@ -16,13 +12,26 @@ async function getNotes() {
   return notes;
 }
 
-async function removeNotes(id) {
-  await Note.deleteOne({ _id: id });
+async function removeNotes(id, owner) {
+  const result = await Note.deleteOne({ _id: id, owner });
+
+  if (result.matchedCount === 0) {
+    throw new Error("No note to delete");
+  }
+
   console.log(chalk.green.inverse(`Note with id ${id} was removed!`));
 }
 
-async function updateNote(noteData) {
-  await Note.updateOne({ _id: noteData.id }, { title: noteData.title });
+async function updateNote(noteData, owner) {
+  const result = await Note.updateOne(
+    { _id: noteData.id, owner },
+    { title: noteData.title }
+  );
+
+  if (result.matchedCount === 0) {
+    throw new Error("No note to edit");
+  }
+
   console.log(chalk.green.inverse(`Note with id ${noteData.id} was updated!`));
 }
